+++ src/migrations/Install.php
@@ -71,6 +71,7 @@ class Install extends Migration
             $this->createTable(ElementCacheRecord::tableName(), [
                 'cacheId' => $this->integer()->notNull(),
                 'elementId' => $this->integer()->notNull(),
+                'expiryDate' => $this->dateTime(),
             ]);
         }